An SPV client must still request all block headers and send a getdata request for every block to decide whether it contains any related transactions. The goal of the bloom filter is to check the number of transactions returned by the merkleblock response to the getdata request.
Without this filter we'd be re-requesting txs from each of our peers, increasing bandwidth consumption considerably. For instance, with 100 peers, half of which relay a tx we don't accept, that might be a 50x bandwidth increase. A flooding attacker attempting to roll-over the
filter using minimum-sized, 60byte, transactions might manage to send 1000/sec if we have fast peers, so we pick 120,000 to give our peers a two minute window to send invs to us.

Signed-off-by: 216k155 <216k155@luxcore.io>